feat: Gate otelcol/static converters behind !slim build tag
